+++ example/lib/src/second_method.dart
@@ -57,10 +57,11 @@ class MyThemedApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'First Method (Themed)',
       builder: (ctx, child) {
+        ScreenUtil.init(ctx);
         return Theme(
           data: ThemeData(
             primarySwatch: Colors.blue,
-            textTheme: TextTheme(bodyText2: TextStyle(fontSize: 30.sp)),
+            textTheme: TextTheme(bodyMedium: TextStyle(fontSize: 30.sp)),
           ),
           child: HomePage(title: 'FlutterScreenUtil Demo'),
         );