+++ example/lib/src_zh/second_method.dart
@@ -56,10 +56,11 @@ class MyThemedApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: '第二种方法（带主题）',
       builder: (ctx, child) {
+        ScreenUtil.init(ctx);
         return Theme(
           data: ThemeData(
             primarySwatch: Colors.blue,
-            textTheme: TextTheme(bodyText2: TextStyle(fontSize: 30.sp)),
+            textTheme: TextTheme(bodyMedium: TextStyle(fontSize: 30.sp)),
           ),
           child: HomePage(title: '第二种方法（带主题）'),
         );