+++ lib/src/size_extension.dart
@@ -14,13 +14,24 @@ extension SizeExtension on num {
   ///[ScreenUtil.radius]
   double get r => ScreenUtil().radius(this);
 
+  ///[ScreenUtil.diagonal]
+  double get dg => ScreenUtil().diagonal(this);
+
+  ///[ScreenUtil.diameter]
+  double get dm => ScreenUtil().diameter(this);
+
   ///[ScreenUtil.setSp]
   double get sp => ScreenUtil().setSp(this);
 
   ///smart size :  it check your value - if it is bigger than your value it will set your value
   ///for example, you have set 16.sm() , if for your screen 16.sp() is bigger than 16 , then it will set 16 not 16.sp()
   ///I think that it is good for save size balance on big sizes of screen
+  double get spMin => min(toDouble(), sp);
+
+  @Deprecated('use spMin instead')
   double get sm => min(toDouble(), sp);
+
+  double get spMax => max(toDouble(), sp);
 
   ///屏幕宽度的倍数
   ///Multiple of screen width
@@ -31,22 +42,38 @@ extension SizeExtension on num {
   double get sh => ScreenUtil().screenHeight * this;
 
   ///[ScreenUtil.setHeight]
-  Widget get verticalSpace => ScreenUtil().setVerticalSpacing(this);
+  SizedBox get verticalSpace => ScreenUtil().setVerticalSpacing(this);
 
   ///[ScreenUtil.setVerticalSpacingFromWidth]
-  Widget get verticalSpaceFromWidth =>
+  SizedBox get verticalSpaceFromWidth =>
       ScreenUtil().setVerticalSpacingFromWidth(this);
 
   ///[ScreenUtil.setWidth]
-  Widget get horizontalSpace => ScreenUtil().setHorizontalSpacing(this);
+  SizedBox get horizontalSpace => ScreenUtil().setHorizontalSpacing(this);
 
   ///[ScreenUtil.radius]
-  Widget get horizontalSpaceRadius =>
+  SizedBox get horizontalSpaceRadius =>
       ScreenUtil().setHorizontalSpacingRadius(this);
 
   ///[ScreenUtil.radius]
-  Widget get verticalSpacingRadius =>
+  SizedBox get verticalSpacingRadius =>
       ScreenUtil().setVerticalSpacingRadius(this);
+
+  ///[ScreenUtil.diameter]
+  SizedBox get horizontalSpaceDiameter =>
+      ScreenUtil().setHorizontalSpacingDiameter(this);
+
+  ///[ScreenUtil.diameter]
+  SizedBox get verticalSpacingDiameter =>
+      ScreenUtil().setVerticalSpacingDiameter(this);
+
+  ///[ScreenUtil.diagonal]
+  SizedBox get horizontalSpaceDiagonal =>
+      ScreenUtil().setHorizontalSpacingDiagonal(this);
+
+  ///[ScreenUtil.diagonal]
+  SizedBox get verticalSpacingDiagonal =>
+      ScreenUtil().setVerticalSpacingDiagonal(this);
 }
 
 extension EdgeInsetsExtension on EdgeInsets {
@@ -57,21 +84,71 @@ extension EdgeInsetsExtension on EdgeInsets {
         right: right.r,
         left: left.r,
       );
+
+  EdgeInsets get dm => copyWith(
+        top: top.dm,
+        bottom: bottom.dm,
+        right: right.dm,
+        left: left.dm,
+      );
+
+  EdgeInsets get dg => copyWith(
+        top: top.dg,
+        bottom: bottom.dg,
+        right: right.dg,
+        left: left.dg,
+      );
+
+  EdgeInsets get w => copyWith(
+        top: top.w,
+        bottom: bottom.w,
+        right: right.w,
+        left: left.w,
+      );
+
+  EdgeInsets get h => copyWith(
+        top: top.h,
+        bottom: bottom.h,
+        right: right.h,
+        left: left.h,
+      );
 }
 
 extension BorderRaduisExtension on BorderRadius {
   /// Creates adapt BorderRadius using r [SizeExtension].
   BorderRadius get r => copyWith(
         bottomLeft: bottomLeft.r,
-        bottomRight: bottomLeft.r,
+        bottomRight: bottomRight.r,
         topLeft: topLeft.r,
         topRight: topRight.r,
+      );
+
+  BorderRadius get w => copyWith(
+        bottomLeft: bottomLeft.w,
+        bottomRight: bottomRight.w,
+        topLeft: topLeft.w,
+        topRight: topRight.w,
+      );
+
+  BorderRadius get h => copyWith(
+        bottomLeft: bottomLeft.h,
+        bottomRight: bottomRight.h,
+        topLeft: topLeft.h,
+        topRight: topRight.h,
       );
 }
 
 extension RaduisExtension on Radius {
   /// Creates adapt Radius using r [SizeExtension].
   Radius get r => Radius.elliptical(x.r, y.r);
+
+  Radius get dm => Radius.elliptical(x.dm, y.dm);
+
+  Radius get dg => Radius.elliptical(x.dg, y.dg);
+
+  Radius get w => Radius.elliptical(x.w, y.w);
+
+  Radius get h => Radius.elliptical(x.h, y.h);
 }
 
 extension BoxConstraintsExtension on BoxConstraints {
@@ -89,5 +166,19 @@ extension BoxConstraintsExtension on BoxConstraints {
         maxWidth: maxWidth.w,
         minHeight: minHeight.h,
         minWidth: minWidth.w,
+      );
+
+  BoxConstraints get w => this.copyWith(
+        maxHeight: maxHeight.w,
+        maxWidth: maxWidth.w,
+        minHeight: minHeight.w,
+        minWidth: minWidth.w,
+      );
+
+  BoxConstraints get h => this.copyWith(
+        maxHeight: maxHeight.h,
+        maxWidth: maxWidth.h,
+        minHeight: minHeight.h,
+        minWidth: minWidth.h,
       );
 }